libstore/globals.cc: Move cgroup detection to libutil
